+++ sdk/resourcemanager/network/armnetwork/constants.go
@@ -11,7 +11,7 @@ package armnetwork
 
 const (
 	moduleName    = "armnetwork"
-	moduleVersion = "v2.2.0"
+	moduleVersion = "v2.3.0"
 )
 
 // Access - Access to be allowed or denied.
@@ -181,6 +181,41 @@ func PossibleApplicationGatewayFirewallModeValues() []ApplicationGatewayFirewall
 	return []ApplicationGatewayFirewallMode{
 		ApplicationGatewayFirewallModeDetection,
 		ApplicationGatewayFirewallModePrevention,
+	}
+}
+
+// ApplicationGatewayFirewallRateLimitDuration - Duration over which Rate Limit policy will be applied. Applies only when
+// ruleType is RateLimitRule.
+type ApplicationGatewayFirewallRateLimitDuration string
+
+const (
+	ApplicationGatewayFirewallRateLimitDurationFiveMins ApplicationGatewayFirewallRateLimitDuration = "FiveMins"
+	ApplicationGatewayFirewallRateLimitDurationOneMin   ApplicationGatewayFirewallRateLimitDuration = "OneMin"
+)
+
+// PossibleApplicationGatewayFirewallRateLimitDurationValues returns the possible values for the ApplicationGatewayFirewallRateLimitDuration const type.
+func PossibleApplicationGatewayFirewallRateLimitDurationValues() []ApplicationGatewayFirewallRateLimitDuration {
+	return []ApplicationGatewayFirewallRateLimitDuration{
+		ApplicationGatewayFirewallRateLimitDurationFiveMins,
+		ApplicationGatewayFirewallRateLimitDurationOneMin,
+	}
+}
+
+// ApplicationGatewayFirewallUserSessionVariable - User Session clause variable.
+type ApplicationGatewayFirewallUserSessionVariable string
+
+const (
+	ApplicationGatewayFirewallUserSessionVariableClientAddr  ApplicationGatewayFirewallUserSessionVariable = "ClientAddr"
+	ApplicationGatewayFirewallUserSessionVariableGeoLocation ApplicationGatewayFirewallUserSessionVariable = "GeoLocation"
+	ApplicationGatewayFirewallUserSessionVariableNone        ApplicationGatewayFirewallUserSessionVariable = "None"
+)
+
+// PossibleApplicationGatewayFirewallUserSessionVariableValues returns the possible values for the ApplicationGatewayFirewallUserSessionVariable const type.
+func PossibleApplicationGatewayFirewallUserSessionVariableValues() []ApplicationGatewayFirewallUserSessionVariable {
+	return []ApplicationGatewayFirewallUserSessionVariable{
+		ApplicationGatewayFirewallUserSessionVariableClientAddr,
+		ApplicationGatewayFirewallUserSessionVariableGeoLocation,
+		ApplicationGatewayFirewallUserSessionVariableNone,
 	}
 }
 
@@ -3929,8 +3964,9 @@ func PossibleWebApplicationFirewallPolicyResourceStateValues() []WebApplicationF
 type WebApplicationFirewallRuleType string
 
 const (
-	WebApplicationFirewallRuleTypeInvalid   WebApplicationFirewallRuleType = "Invalid"
-	WebApplicationFirewallRuleTypeMatchRule WebApplicationFirewallRuleType = "MatchRule"
+	WebApplicationFirewallRuleTypeInvalid       WebApplicationFirewallRuleType = "Invalid"
+	WebApplicationFirewallRuleTypeMatchRule     WebApplicationFirewallRuleType = "MatchRule"
+	WebApplicationFirewallRuleTypeRateLimitRule WebApplicationFirewallRuleType = "RateLimitRule"
 )
 
 // PossibleWebApplicationFirewallRuleTypeValues returns the possible values for the WebApplicationFirewallRuleType const type.
@@ -3938,6 +3974,7 @@ func PossibleWebApplicationFirewallRuleTypeValues() []WebApplicationFirewallRule
 	return []WebApplicationFirewallRuleType{
 		WebApplicationFirewallRuleTypeInvalid,
 		WebApplicationFirewallRuleTypeMatchRule,
+		WebApplicationFirewallRuleTypeRateLimitRule,
 	}
 }